+++ packages/skeleton-react/test/components/app-bar/app-bar.tsx
@@ -1,0 +1,13 @@
+import { AppBar } from '@/index';
+
+export default function AppBarTest() {
+	return (
+		<AppBar data-testid="root">
+			<AppBar.Toolbar data-testid="toolbar">
+				<AppBar.Lead data-testid="lead" />
+				<AppBar.Headline data-testid="headline" />
+				<AppBar.Trail data-testid="trail" />
+			</AppBar.Toolbar>
+		</AppBar>
+	);
+}